refactor(core): extract coordinator module (3/3)
Move the coordinator subsystem out of orchestrator.ts into coordinator.ts:
goal-decomposition parsing, coordinator prompt construction (system prompt,
roster, output format, decomposition), the synthesis pass, and task-spec loading.

The parse helpers are a pure move. The former private class methods become free
functions; two take the orchestrator config explicitly (buildCoordinatorBaseConfig,
runCoordinatorSynthesis) instead of this.config. Prompt string literals are
byte-for-byte unchanged (only code indentation shifts as methods de-indent to
module scope). One type follow-on: runCoordinatorSynthesis gains a
`?? DEFAULT_MODEL` guard because OrchestratorConfig.defaultModel is optional where
the resolved class field was not; it is dead at runtime (callers pass the resolved
config) and mirrors the same guard already in task-execution's executeQueue.

orchestrator.ts drops from 2213 to 1683 lines (57% below the original 3953). Also
updates the AGENTS.md layer map.

Verified with npm run lint, 1385 passing tests, zero diff under tests/, a
byte-identical parse block, all 87 single-quoted and 17 backtick prompt-string
literals identical, and a dist/index.d.ts identical to the pre-split build.

Third of three sequential pure-move PRs; builds on the merged #400 and #397.
